flatten filter into page
